+++ src/body.js
@@ -415,11 +415,9 @@ export function clone(instance) {
  *
  * This function assumes that instance.body is present.
  *
- * @param   Mixed  instance  Response or Request instance
+ * @param   Mixed  instance  Any options.body input
  */
 export function extractContentType(body) {
-	// istanbul ignore if: Currently, because of a guard in Request, body
-	// can never be null. Included here for completeness.
 	if (body === null) {
 		// body is null
 		return null;
@@ -466,7 +464,6 @@ export function extractContentType(body) {
 export function getTotalBytes(instance) {
 	const {body} = instance;
 
-	// istanbul ignore if: included for completion
 	if (body === null) {
 		// body is null
 		return 0;
@@ -484,7 +481,7 @@ export function getTotalBytes(instance) {
 		return null;
 	} else {
 		// body is stream
-		return instance.size || null;
+		return null;
 	}
 }